requirements.txt
django == 2.2.4
django-cors-headers == 3.2
django_heroku == 0.3.1
whitenoise == 5.0.1
psycopg2 == 2.8.3
gunicorn == 19.9.0
psycopg2-binary == 2.8.3
django_compressor == 2.3
django-libsass == 0.7
django-markdown-deux == 1.0.5
django_taggit == 1.2.0
django-crispy-forms==1.8.0
django-bootstrap4==1.0.1
djangorestframework == 3.11
scikit-learn == 0.22.2.post1
pandas == 1.0.1
numpy == 1.18.1
django-tables2==2.1.1
django-filter==2.2.0
django-bootstrap3==11.1.0
